Initial commit
A relatively minimal setup for Android development. The "relative"
parts:

- A subproject is included. This isn't strictly necessary but a lot of
tools depend on this being the standard form.q

- The MainActivity.kt file is contained in a directory 4 levels
deep. Again, not strictly necessary as this could live at the top
level just fine. But a lot of tools expect it to be somewhere
deep. There may be some tools that expect it to also be at the bottom
of a hierarchy named after the class itself but I haven't found that
to be essential.